Fix to write the hex byte value.

+++ src/bootstrap_linker/writeLine.cpp
@@ -29,7 +29,7 @@ void bootstrap_linker::writeLine(std::ostream& out)
 
     for (size_t i = 0; i < 32 && !bytes.empty(); ++i)
     {
-        out << hex << setw(2) << setfill('0') << bytes.front();
+        out << hex << setw(2) << setfill('0') << (int)bytes.front();
         bytes.pop_front();
         ++deque_address;
     }
